requirements.txt
python-dateutil==2.8.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.2
